+++ packages/utils/src/index.ts
@@ -1,8 +1,17 @@
-export type { FormatTimeOptions, FormattedTime, TimeUnit } from './time'
+export { getPasswordStrength } from './password'
+export type {
+	FormatTimeOptions,
+	FormattedTime,
+	GreetingKey,
+	SimpleGreetingKey,
+	TimeUnit,
+} from './time'
 export {
 	formatRateLimitTime,
 	formatTime,
 	formatTimeWithI18n,
+	getGreetingKey,
+	getSimpleGreetingKey,
 	getTimeUnitKey,
 	getTzOffset,
 } from './time'