+++ apps/native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # native
 
+## 0.2.4
+
+### Patch Changes
+
+- feat: ✨ make ui component more accessible
+- Updated dependencies
+  - @folionote/locales@0.2.4
+  - @folionote/utils@0.2.4
+  - @folionote/api@0.2.4
+  - @folionote/constants@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes