+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @folionote/constants
 
+## 0.2.4
+
+### Patch Changes
+
+- feat: ✨ make ui component more accessible
+- Updated dependencies
+  - @folionote/locales@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes